Handle Compilation Threads at Checkpoint/Restore
In order to ensure that a checkpoing/restore does not happen in the
middle of a compilation, when the J9HOOK_VM_PREPARING_FOR_CHECKPOINT
hook is called, all compilation threads are slated to be suspended.
Any existing compilations are allowed to complete. At restore time,
the compilation threads are resumed.

One difference between suspension at checkpoint time vs at any other
point is that the compilation queue is not purged. The reason for
this is that we want execution to continue as normal on restore.

Normal Execution (steps 5&6 can be interchanged)
================================================
1. Hook Thread acquires Comp Monitor
2. Hook Thread signals Comp Threads to suspend
3. Hook Thread acquires CR Monitor, releases Comp Monitor, and waits on
   CR Monitor
4. Comp Threads acquire Comp Monitor and CompThread Monitor, and change
   state to COMPTHREAD_SUSPENDED
5. Comp Threads acquire CR Monitor, call notifyAll, and release CR
   Monitor
6. Hook Thread wakes up with CR Monitor in hand, releases CR Monitor, and
   blocks on Comp Monitor
7. Comp Threads release Comp Monitor and wait on CompThread Monitor
8. Hook Thread acquires Comp Monitor, ensures state has changed to
   COMPTHREAD_SUSPENDED for the current compInfoPT
9. Hook Thread checks the next compInfoPT, waiting on the CR Monitor if
   needed (it will release the Comp Monitor prior to waiting)
10. Hook Thread releases Comp Monitor, returns from the hook

JIT Dump
========
- Crash on application thread:
   - Hook Thread running prepareForCheckpoint will run as normal
      - Comp Threads will suspend themselves
      - Hook Thread will wait till all Comp Threads are suspended
      - Hook Thread will return from the jit hook
      - VM will need to ensure it doesn't invoke criu API

- Crash on compilation thread:
   - Crashing Comp Thread will return to the VM and terminate process
     (https://github.com/eclipse-openj9/openj9/blob/500e0a26e2c5be6ead0f838495ae8d8cc34821e1/runtime/compiler/control/CompilationThread.cpp#L3442-L3447)
   - Possible scenarios for Hook Thread:
      1. Hook Thread will try to acquire the Comp Monitor to signal Comp
         Threads to suspend but will end up blocking until the JVM
         terminates
      2. Hook Thread will manage to signal Comp Threads to suspend, but
         will remain waiting on the CR Monitor for the crashed Comp
         Thread to suspend until the JVM terminates
   - VM will need to ensure it doesn't invoke criu API

Normal Shutdown
===============
- Scenario 1
   1. Hook Thread waits on CR Monitor
   2. Comp Thread goes to suspend, already has Comp Monitor in hand
   3. Shutdown Thread blocks on the Comp Monitor
   4. Comp Thread updates state to COMPTHREAD_SUSPENDED, acquires CR
      Monitor, and calls notifyAll
   5. Comp Thread releases CR Monitor, releases Comp Monitor, and waits
      on CompThread Monitor
   6. Hook Thread wakes up with CR Monitor in hand, releases CR Monitor,
      acquires Comp Monitor and continues on to next thread
   7. Hook Thread checks state and moves onto the next threads possibly
      finding them all suspended.
   8. Hook Thread release Comp Monitor and returns from hook
   9. Shutdown Thread goes through the sequence of stopping all threads

- Scenario 2
   Repeat steps 1-5 from Scenario 1
   6. Shutdown Thread acquires Comp Monitor
   7. Hook Thread wakes up with CR Monitor in hand, releases CR Monitor
      and blocks on the Comp Monitor
   8. Shutdown Thread sets checkpoint to be interrupted flag.
      Additionally (though irrelevant in this scenario) it also
      acquires the CR Monitor, calls notify, and releases CR Monitor.
   9. Shutdown Thread finishes stopping all threads, releases Comp
      Monitor
   10. Hook Thread acquires Comp Monitor, sees that the checkpoint should
       be interrupted
   11. Hook Thread breaks out of the loops, releases Comp Monitor, and
       returns from hook

- Scenario 3
   1. Hook Thread waits on CR Monitor
   2. Shutdown Thread acquires Comp Monitor
   3. Comp Thread blocks on the Comp Monitor in order to (eventually)
      suspend itself
   4. Shutdown Thread sets checkpoint to be interrupted flag, acquires CR
      Monitor, calls notify, and releases CR Monitor
   5. Shutdown Thread goes about the task of stopping Comp Threads
   6. Hook Thread wakes up with CR Monitor in hand, releases CR Monitor
      and blocks on the Comp Monitor
   7. Shutdown Thread finishes stopping all threads, releases Comp
      Monitor (steps 5 & 6 can be interchanged with the same following
      steps)
   8. Hook Thread acquires Comp Monitor, sees that the checkpoint should
      be interrupted
   9. Hook Thread breaks out of the loops, releases Comp Monitor, and
      returns from hook

It should be noted that when the Hook Thread runs prepareForCheckpoint,
either it will succeed in signalling the Comp Threads to suspend, or it
will wait indefinitely on the Comp Monitor (in the case of a JIT Dump) or
it will abort after acquring the Comp Monitor if the Shutdown Thread
already finished its task. Once the Hook Thread reaches the point where
it waits on the CR Monitor, the scenarioes above can occur.

Shutdown during JIT Dump
========================
- Crash on application thread
   - same as Normal Shutdown

- Crash on compilation thread
   - In all of the above scenarios:
      - Hook Thread remains waiting on CR Monitor
      - Shutdown Thread blocks on the Comp Monitor
      - Comp Thread returns to VM and terminates the process

+++ runtime/compiler/control/CompilationThread.cpp
@@ -2449,13 +2449,8 @@ void TR::CompilationInfoPerThread::suspendCompilationThread()
 // This method may be called when we want all JIT activity to stop temporarily
 // It is called normally by a java thread
 //----------------------------------------------------------------------------
-void TR::CompilationInfo::suspendCompilationThread()
+void TR::CompilationInfo::suspendCompilationThread(bool purgeCompQueue)
    {
-   // if we compile on application thread, there is no compilation
-   // request queue and there is no compilation thread
-   // There can only be one request that is in progress. All the others
-   // will see the SUSPENDED state when they get the application monitor
-
    TR_ASSERT(_compilationMonitor, "Must have a compilation queue monitor\n");
    J9JavaVM   *vm       = _jitConfig->javaVM;
    J9VMThread *vmThread = vm->internalVMFunctions->currentVMThread(vm);
@@ -2484,7 +2479,7 @@ void TR::CompilationInfo::suspendCompilationThread()
             }
          }
       }
-   if (stoppedOneCompilationThread)
+   if (stoppedOneCompilationThread && purgeCompQueue)
       purgeMethodQueue(compilationSuspended);
    TR_ASSERT(_numCompThreadsActive == 0, "We must have all compilation threads suspended at this point\n");
    releaseCompMonitor(vmThread);
@@ -2617,6 +2612,93 @@ void TR::CompilationInfo::resumeCompilationThread()
    releaseCompMonitor(vmThread);
    }
 
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+void TR::CompilationInfo::prepareForCheckpoint()
+   {
+   J9JavaVM   *vm       = _jitConfig->javaVM;
+   J9VMThread *vmThread = vm->internalVMFunctions->currentVMThread(vm);
+
+   {
+   OMR::CriticalSection suspendCompThreadsForCheckpoint(getCompilationMonitor());
+
+   if (shouldCheckpointBeInterrupted())
+      return;
+
+   TR_ASSERT_FATAL(!isCheckpointInProgress(), "Checkpoint already in progress!\n");
+
+   /* Set the checkpoint in progress status. */
+   setCheckpointInProgress();
+
+   /* Inform compilation threads to suspend themselves. The compilation
+    * threads will complete any in-progress compilations first.
+    */
+   bool purgeMethodQueue = false;
+   suspendCompilationThread(purgeMethodQueue);
+
+   /* Wait until all compilation threads are suspended. */
+   for (int32_t i = 0; i < getNumTotalCompilationThreads(); i++)
+      {
+      TR::CompilationInfoPerThread *curCompThreadInfoPT = _arrayOfCompilationInfoPerThread[i];
+      if (!curCompThreadInfoPT->isDiagnosticThread())
+         {
+         do
+            {
+            /* Determine whether to wait on the CR Monitor.
+             *
+             * IMPORTANT: There should be no return, or C++ exception thrown, after
+             *            releasing the Comp Monitor below until it is re-acquired.
+             *            The Comp Monitor was acquired in an OMR::CriticalSection
+             *            object; a return, or thrown exception, will destruct this
+             *            object as part leaving the scope, or stack unwinding, which
+             *            will attempt to release the monitor in its destructor.
+             */
+            if (!shouldCheckpointBeInterrupted()
+                && curCompThreadInfoPT->getCompilationThreadState() != COMPTHREAD_SUSPENDED)
+               {
+               /* Acquire the CR monitor */
+               acquireCRMonitor();
+
+               /* Release the Comp Monitor before waiting */
+               releaseCompMonitor(vmThread);
+
+               /* Wait until notified */
+               waitOnCRMonitor();
+
+               /* Release CR Monitor and re-acquire the Comp Monitor */
+               releaseCRMonitor();
+               acquireCompMonitor(vmThread);
+               }
+            else
+               {
+               break;
+               }
+            }
+         while(true);
+         }
+
+      /* Stop cycling through the threads if checkpointing should be interrupted. */
+      if (shouldCheckpointBeInterrupted())
+         break;
+      }
+   }
+
+   }
+
+void TR::CompilationInfo::prepareForRestore()
+   {
+
+   {
+   OMR::CriticalSection resumeCompThreadsForRestore(getCompilationMonitor());
+
+   /* Reset the checkpoint in progress flag. */
+   resetCheckpointInProgress();
+
+   /* Resume suspended compilation threads. */
+   resumeCompilationThread();
+   }
+
+   }
+#endif // #if defined(J9VM_OPT_CRIU_SUPPORT)
 
 int TR::CompilationInfo::computeCompilationThreadPriority(J9JavaVM *vm)
    {
@@ -3428,6 +3510,18 @@ void TR::CompilationInfo::stopCompilationThreads()
 #endif
    acquireCompMonitor(vmThread);
 
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+   // Interrupt any checkpoint in progress and notify
+   // the thread waiting in the checkpoint hook.
+   if (isCheckpointInProgress())
+      {
+      interruptCheckpoint();
+      acquireCRMonitor();
+      getCRMonitor()->notifyAll();
+      releaseCRMonitor();
+      }
+#endif
+
    // Cycle through all non-diagnostic threads and stop them
    for (int32_t i = 0; i < getNumTotalCompilationThreads(); i++)
       {
@@ -3893,13 +3987,34 @@ TR::CompilationInfoPerThread::doSuspend()
    {
    _compInfo.setSuspendThreadDueToLowPhysicalMemory(false);
    getCompThreadMonitor()->enter();
+
+   // Set the new state
    setCompilationThreadState(COMPTHREAD_SUSPENDED);
-   _compInfo.releaseCompMonitor(getCompilationThread());   // release the queue monitor before waiting
+
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+   // Notify the thread waiting in the checkpoint hook.
+   if (_compInfo.isCheckpointInProgress())
+      {
+      _compInfo.acquireCRMonitor();
+      _compInfo.getCRMonitor()->notifyAll();
+      _compInfo.releaseCRMonitor();
+      }
+#endif
+
+   // release the queue monitor before waiting
+   _compInfo.releaseCompMonitor(getCompilationThread());
+
    setLastTimeThreadWasSuspended(_compInfo.getPersistentInfo()->getElapsedTime());
    setVMThreadNameWithFlag(getCompilationThread(), getCompilationThread(), getSuspendedThreadName(), 1);
-   getCompThreadMonitor()->wait(); // wait here until someone notifies us
+
+   // wait here until someone notifies us
+   getCompThreadMonitor()->wait();
+
    setVMThreadNameWithFlag(getCompilationThread(), getCompilationThread(), getActiveThreadName(), 1);
+
    getCompThreadMonitor()->exit();
+
+   // reacquire the queue monitor
    _compInfo.acquireCompMonitor(getCompilationThread());
    }
